create command: add dir options
